Work for #7181 - Cover mobile view - renamed some files
