+++ src/defaultV2-theme/blocks/header.scss
