+++ src/defaultV2-theme/defaultV2.fontless.scss
@@ -47,7 +47,7 @@
 @import "blocks/sd-progress-toc.scss";
 @import "blocks/sd-list.scss";
 @import "blocks/sd-timer.scss";
-@import "blocks/cover.scss";
+@import "blocks/header.scss";
 @import "blocks/sd-loading-indicator.scss";
 @import "../components-container.scss";
 @import "../signaturepad.scss";
